ajout commande avec produits

+++ src/main/java/com/msr/cg/afrimeta/commande/dto/CommandeDto.java
@@ -18,7 +18,7 @@ public record CommandeDto(
 
         @NotNull
     @Column(name = "commande_total")
-    int commandeTotal,
+    double commandeTotal,
 
         @NotNull
     String adresse,
